fixup some stuff that broke when i was moving stuff around
